% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anytime-measure.R
\name{anytime_measure}
\alias{anytime_measure}
\title{A measure of anytime performance}
\usage{
anytime_measure(
  data,
  tvar,
  qvar,
  rvar,
  avar,
  tlim = NULL,
  qlim = NULL,
  maximizing = TRUE
)
}
\arguments{
\item{data}{input data frame}

\item{tvar}{variable with the time data}

\item{qvar}{variable with the quality data}

\item{rvar}{variable with the run id}

\item{avar}{variable with the algorithm id}

\item{tlim}{limits for time, vector of length 2, or NULL to use the
limits from the data frame}

\item{qlim}{limits for quality, vector of length 2, or NULL to use
the limits from the data frame}

\item{maximizing}{boolean denoting whether the quality is being
maximized or minimized}
}
\description{
This functions provides a measure of anytime performance as described
in "A. D. Jesus, A. Liefooghe, B. Derbel, L. Paquete. Algorithm
Selection of Anytime Algorithms. Proceedings of the 2020 Genetic and
Evolutionary Computation Conference (GECCO 2020), 850-858, 2020."
}
\examples{
anytime_measure(traces_small, time, quality, run, algo)
}
