% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confine.R
\name{confine}
\alias{confine}
\title{Confine the performance traces to a bounding box.}
\usage{
confine(
  data,
  tvar,
  qvar,
  ...,
  tlim = NULL,
  qlim = NULL,
  qmax = TRUE,
  extend = TRUE
)
}
\arguments{
\item{data}{data frame to preprocess}

\item{tvar}{variable that gives the time column in the data}

\item{qvar}{variable that gives the quality column in the data}

\item{...}{grouping variables for the performance traces}

\item{tlim}{numeric vector of length 2 denoting the limits for time,
or \code{NULL} to use the limits of the data}

\item{qlim}{numeric vector of length 2 denoting the limits for
quality, or \code{NULL} to use the limits of the data}

\item{qmax}{boolean denoting whether the quality measure is being
maximized or minimized}

\item{extend}{boolean denoting whether or not to add a fake
observation, for each performance trace, at the limits}
}
\description{
\code{confine()} confines the performance traces to a given bounding box.
Also, it optionally extends the performance traces until the limits
of bounding box. This method takes into account whether the quality
is being maximized or minimized.
}
\examples{
require(dplyr)

confine(traces_small, time, quality, algo, run,
        tlim = c(0, 5), qlim = c(0, 1), extend = TRUE)

traces_small \%>\% mutate(quality = 1 - quality) \%>\%
confine(time, quality, algo, run,
        tlim = c(0, 5), qlim = c(0, 1), qmax = FALSE, extend = TRUE)
}
