% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganytime-epp.R
\name{gganytime_epp}
\alias{gganytime_epp}
\title{Plot the empirical performance profile of an anytime algorithm.}
\usage{
gganytime_epp(
  data,
  tvar,
  qvar,
  rvar,
  avar,
  algo = NULL,
  tlim = NULL,
  qlim = NULL,
  maximizing = T
)
}
\arguments{
\item{data}{data frame with the anytime data}

\item{tvar}{time variable in the data}

\item{qvar}{quality variable in the data}

\item{rvar}{run id variable in the data}

\item{avar}{algorithm id variable in the data}

\item{algo}{which algorithm to choose for plotting, if there is only
a single one in the data you may ignore this}

\item{tlim}{limits for time}

\item{qlim}{limits for quality}

\item{maximizing}{whether the quality is being maximized or minimized}
}
\description{
Plot the empirical performance profile of an anytime algorithm.
}
\examples{
gganytime_epp(traces_small, time, quality, run, algo, 1)
}
