% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganytime-epp.R
\name{gganytime_epp_min}
\alias{gganytime_epp_min}
\title{Plots the minimum positive difference between the empirical
performance profile of an algorithm and all others}
\usage{
gganytime_epp_min(
  data,
  tvar,
  qvar,
  rvar,
  avar,
  tlim = NULL,
  qlim = NULL,
  maximizing = T
)
}
\arguments{
\item{data}{data frame with the anytime data}

\item{tvar}{time variable in the data}

\item{qvar}{quality variable in the data}

\item{rvar}{run id variable in the data}

\item{avar}{algorithm id variable in the data}

\item{tlim}{limits for time}

\item{qlim}{limits for quality}

\item{maximizing}{whether the quality is being maximized or minimized}
}
\description{
Plots the minimum positive difference between the empirical
performance profile of an algorithm and all others
}
\examples{
gganytime_epp_min(traces_small, time, quality, run, algo)
}
