% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganytime-epp.R
\name{gganytime_epp_pair}
\alias{gganytime_epp_pair}
\title{Plots pairwise differences between the empirical performance profiles.}
\usage{
gganytime_epp_pair(
  data,
  tvar,
  qvar,
  rvar,
  avar,
  tlim = NULL,
  qlim = NULL,
  maximizing = T
)
}
\arguments{
\item{data}{data frame with the anytime data}

\item{tvar}{time variable in the data}

\item{qvar}{quality variable in the data}

\item{rvar}{run id variable in the data}

\item{avar}{algorithm id variable in the data}

\item{tlim}{limits for time}

\item{qlim}{limits for quality}

\item{maximizing}{whether the quality is being maximized or minimized}
}
\description{
Plots pairwise differences between the empirical performance profiles.
}
\examples{
gganytime_epp_pair(traces_small, time, quality, run, algo)
}
