% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganytime-eqd.R
\name{gganytime_eqd}
\alias{gganytime_eqd}
\title{Plots the empirical time distribution for a specific quality}
\usage{
gganytime_eqd(
  data,
  tvar,
  qvar,
  time,
  tgroup,
  agroup,
  aes = aes(),
  qlim = NULL,
  maximizing = TRUE,
  extend = TRUE,
  line = "step"
)
}
\arguments{
\item{data}{dataset}

\item{tvar}{variable with time information}

\item{qvar}{variable with quality information}

\item{time}{a numeric value denoting the fixed time}

\item{tgroup}{trace grouping variables (use c(...))}

\item{agroup}{aesthetics grouping variables (use c(...))}

\item{aes}{aesthetics mapping}

\item{qlim}{limits for time}

\item{maximizing}{boolean denoting whether or not the quality measure
is being maximized}

\item{extend}{boolean denoting whether or not to extend the lines to
infinity}

\item{line}{one of "step" or "line"}
}
\description{
\code{gganytime_eqd()} calculates the empirical time distribution,
which gives the minimum time needed to achieve a certain quality.
}
\examples{
require(ggplot2)
gganytime_eqd(traces_large, time, quality, 5, c(run, algo), c(algo),
              aes = aes(colour = algo))
}
