% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganytime-etd.R
\name{gganytime_etd}
\alias{gganytime_etd}
\title{Plots the empirical time distribution for a specific quality}
\usage{
gganytime_etd(
  data,
  tvar,
  qvar,
  quality,
  tgroup,
  agroup,
  aes = aes(),
  tlim = NULL,
  maximizing = TRUE,
  extend = TRUE,
  line = "step"
)
}
\arguments{
\item{data}{dataset}

\item{tvar}{variable with time information}

\item{qvar}{variable with quality information}

\item{quality}{a numeric value denoting the fixed quality}

\item{tgroup}{trace group (use c(...))}

\item{agroup}{aesthetics group (use c(...))}

\item{aes}{aesthetics mapping}

\item{tlim}{limits for time}

\item{maximizing}{boolean denoting whether or not the quality measure
is being maximized}

\item{extend}{boolean denoting whether or not to extend the lines to
the limits}

\item{line}{one of "step" or "line"}
}
\description{
\code{gganytime_etd()} calculates the empirical time distribution,
which gives the minimum time needed to achieve a certain quality.
}
\examples{
require(ggplot2)
gganytime_etd(traces_large, time, quality, 0.5, c(run, algo), c(algo),
              aes = aes(colour = algo))
}
