% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{iso8601}
\alias{iso8601}
\title{Format a Datatime object according to ISO 8601}
\usage{
iso8601(pt)
}
\arguments{
\item{pt}{A Datetime object}
}
\value{
A character object
}
\description{
ISO 8601 is a standard for date and time representation covering the
date and time (usually excluding possible fractional seconds)..
It is described in TODO
}
\examples{
iso8601(anytime("2016-09-01 10:11:12.123456"))
iso8601(anydate("2016-Sep-01"))
}

