% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{rfc2822}
\alias{rfc2822}
\title{Format a Datatime object according to RFC 2822}
\usage{
rfc2822(pt)
}
\arguments{
\item{pt}{A Datetime object}
}
\value{
A character object
}
\description{
RFC 3339 is a standard for date and time representation covering
the weekday, date, time (including possible fractional seconds) and
a timezone offset.  It is described in TODO
}
\examples{
rfc2822(anytime("2016-09-01 10:11:12.123456"))
rfc2822(anydate("2016-Sep-01"))
}

