% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{rfc3339}
\alias{rfc3339}
\title{Format a Datatime object according to RFC 3339}
\usage{
rfc3339(pt)
}
\arguments{
\item{pt}{A Datetime object}
}
\value{
A character object
}
\description{
RFC 3339 is a standard for date and time representation covering the
date, time (including possible fractional seconds) and a timezone offset.
It is described in TODO
}
\examples{
rfc3339(anytime("2016-09-01 10:11:12.123456"))
rfc3339(anydate("2016-Sep-01"))
}

