% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traces_large}
\alias{traces_large}
\alias{traces_small}
\alias{nonmonotonic_traces_large}
\alias{nonmonotonic_traces_small}
\title{Generated performance traces.}
\format{
A data frame with 4 variables:
\describe{
\item{time}{execution time in seconds (0--10)}
\item{quality}{quality of the solution (0--1)}
\item{run}{id for the run (factor)}
\item{algo}{id for the algorithm (factor)}
}

An object of class \code{data.frame} with 90 rows and 4 columns.

An object of class \code{data.frame} with 900 rows and 4 columns.

An object of class \code{data.frame} with 90 rows and 4 columns.
}
\usage{
traces_large

traces_small

nonmonotonic_traces_large

nonmonotonic_traces_small
}
\description{
A dataset containing generated performance traces for three
algorithms. \code{traces_large} contains 30 monotonic performance traces
for each algorithm, \code{traces_small} contains only the first 3
performance traces for each algorithm. \code{nonmonotonic_traces_large}
contains 30 nonmonotonic performance traces for each algorithm,
\code{nonmonotonic_traces_small} contains only the first 3 performance
traces for each algorithm. The variables are as follows:
}
\keyword{datasets}
