% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFormats}
\alias{addFormat}
\alias{getFormats}
\title{Functions to retrieve (or set) formats used for parsing dates.}
\usage{
getFormats()

addFormat(fmt)
}
\arguments{
\item{fmt}{A character value in the form understood by Boost date_time}
}
\value{
Nothing in the case of \code{addFormat}; a character vector of
formats in the case of \code{getFormats}
}
\description{
The time and date parsing and conversion relies on trying a (given
and fixed) number of timeformats. The format used is the one employed
by the underlying implementation of the Boost date_time library.
}
\examples{
  getFormats()
  addFormat("\%d \%b \%y")   # two-digit date [not recommended], textual month
}
\author{
Dirk Eddelbuettel
}
\seealso{
\code{\link{anytime-package}} and references therein
}

